Reorganise utils files
